migrations/20221129_01_7b5bV.py
"""
sample
"""

from yoyo import step

__depends__ = {}

steps = [
    step("""CREATE TABLE documents (
    id serial,
    rubrics text[],
    text text,
    created_date date
    );""",
         """DROP TABLE documents;""")
]
