requirements.txt
envparse==0.2.0
fastapi==0.88.0
elasticsearch==8.5.2
asyncpg==0.27.0
uvicorn==0.20.0
aiohttp==3.8.3
pydantic==1.10.2
pytest==7.2.0
httpx==0.23.1
yoyo-migrations==8.1.0
psycopg2==2.9.5
